+++ components/src/hardware-sim/ProtocolDeck/utils/index.ts
@@ -1,2 +1,2 @@
-export * from './getLabwareOnDeck'
+export * from './getLabwareInSlots'
 export * from './getStandardDeckViewLayerBlockList'