+++ protocol-designer/src/components/modals/CreateFileWizard/PipetteTypeTile.tsx
@@ -10,8 +10,6 @@ import {
   ALIGN_CENTER,
   PrimaryButton,
   JUSTIFY_SPACE_BETWEEN,
-  LEFT,
-  RIGHT,
   InstrumentDiagram,
 } from '@opentrons/components'
 import {
@@ -21,6 +19,8 @@ import {
   OT3_PIPETTES,
   getAllPipetteNames,
   getPipetteNameSpecs,
+  LEFT,
+  RIGHT,
 } from '@opentrons/shared-data'
 
 import { i18n } from '../../../localization'