+++ protocol-designer/src/components/modules/FlexSlotMap.tsx
@@ -16,8 +16,6 @@ import {
   SPACING,
 } from '@opentrons/components'
 
-import type { Cutout } from '@opentrons/shared-data'
-
 const X_ADJUSTMENT_LEFT_SIDE = -101.5
 const X_ADJUSTMENT = -17
 const X_DIMENSION_MIDDLE_SLOTS = 160.3
@@ -49,7 +47,7 @@ export function FlexSlotMap(props: FlexSlotMapProps): JSX.Element {
       {deckDef.locations.cutouts.map(cutout => (
         <SingleSlotFixture
           key={cutout.id}
-          cutoutId={cutout.id as Cutout}
+          cutoutId={cutout.id}
           deckDefinition={deckDef}
           slotClipColor={COLORS.transparent}
           fixtureBaseColor={COLORS.light1}