+++ protocol-designer/src/step-forms/test/createPresavedStepForm.test.ts
@@ -49,6 +49,9 @@ beforeEach(() => {
         labwareOnMagModule,
       },
     },
+    additionalEquipmentEntities: {
+      mockTrash: { name: 'trashBin', id: 'mockTrash', location: 'A3' },
+    },
     savedStepForms: {},
     orderedStepIds: [],
     initialDeckSetup: {
@@ -122,7 +125,7 @@ describe('createPresavedStepForm', () => {
       })
     })
   })
-  it(`should call handleFormChange with a default pipette for "moveLiquid" step`, () => {
+  it(`should call handleFormChange with a default pipette and drop tip location for "moveLiquid" step`, () => {
     const args = { ...defaultArgs, stepType: 'moveLiquid' }
     expect(createPresavedStepForm(args)).toEqual({
       id: stepId,
@@ -130,7 +133,7 @@ describe('createPresavedStepForm', () => {
       nozzles: null,
       stepType: 'moveLiquid',
       // default fields
-      dropTip_location: null,
+      dropTip_location: 'mockTrash',
       aspirate_airGap_checkbox: false,
       aspirate_airGap_volume: '1',
       aspirate_delay_checkbox: false,
@@ -177,7 +180,7 @@ describe('createPresavedStepForm', () => {
     })
   })
   describe('mix step', () => {
-    it('should call handleFormChange with a default pipette for mix step', () => {
+    it('should call handleFormChange with a default pipette and drop tip location for mix step', () => {
       const args = { ...defaultArgs, stepType: 'mix' }
       expect(createPresavedStepForm(args)).toEqual({
         id: stepId,
@@ -186,7 +189,7 @@ describe('createPresavedStepForm', () => {
         // default fields
         labware: null,
         nozzles: null,
-        dropTip_location: null,
+        dropTip_location: 'mockTrash',
         wells: [],
         aspirate_delay_checkbox: false,
         aspirate_delay_seconds: `${DEFAULT_DELAY_SECONDS}`,