+++ protocol-designer/src/utils/index.ts
@@ -19,6 +19,7 @@ import type {
   LabwareEntities,
   PipetteEntities,
 } from '@opentrons/step-generation'
+import { INTERACTIVE_WELL_DATA_ATTRIBUTE } from '@opentrons/components/src/hardware-sim/Labware/labwareInternals/Well'
 
 export const registerSelectors: (arg0: any) => void =
   process.env.NODE_ENV === 'development'
@@ -43,10 +44,7 @@ export function clientRectToBoundingRect(rect: ClientRect): BoundingRect {
     height: rect.height,
   }
 }
-export const getCollidingWells = (
-  rectPositions: GenericRect,
-  selectableClassname: string
-): WellGroup => {
+export const getCollidingWells = (rectPositions: GenericRect): WellGroup => {
   // Returns set of selected wells under a collision rect
   const { x0, y0, x1, y1 } = rectPositions
   const selectionBoundingRect = {
@@ -56,9 +54,10 @@ export const getCollidingWells = (
     height: Math.abs(y1 - y0),
   }
   // NOTE: querySelectorAll returns a NodeList, so you need to unpack it as an Array to do .filter
-  // @ts-expect-error(sa, 2021-6-21): there is no option to query by class selector in HTMLElementTagNameMap (see type of querySelectorAll)
   const selectableElems: HTMLElement[] = [
-    ...document.querySelectorAll('.' + selectableClassname),
+    ...document.querySelectorAll<HTMLElement>(
+      `[${INTERACTIVE_WELL_DATA_ATTRIBUTE}]`
+    ),
   ]
   const collidedElems = selectableElems.filter((selectableElem, i) =>
     rectCollision(
@@ -68,11 +67,11 @@ export const getCollidingWells = (
   )
   const collidedWellData = collidedElems.reduce(
     (acc: WellGroup, elem): WellGroup => {
-      // TODO IMMEDIATELY no magic string 'wellname'
-      if ('wellname' in elem.dataset) {
+      if (
+        INTERACTIVE_WELL_DATA_ATTRIBUTE.replace('data-', '') in elem.dataset
+      ) {
         const wellName = elem.dataset.wellname
-        // @ts-expect-error(sa, 2021-6-21): wellName might be undefined
-        return { ...acc, [wellName]: null }
+        return wellName != null ? { ...acc, [wellName]: null } : acc
       }
 
       return acc
@@ -81,7 +80,6 @@ export const getCollidingWells = (
   )
   return collidedWellData
 }
-// TODO IMMEDIATELY use where appropriate
 export const arrayToWellGroup = (w: string[]): WellGroup =>
   w.reduce((acc, wellName) => ({ ...acc, [wellName]: null }), {})
 // cross-PD memoization of well set utils